Move make_response to _docker
